+++ packages/secrets/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe Secrets SDK package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Updated the `adler` dependency to `adler2` for technical currency on MacOS. [#2600](https://github.com/zowe/zowe-cli/pull/2600)
+
 ## `8.24.2`
 
 - BugFix: Fixed an issue where the buffer for error handling on Windows was passed by value instead of by reference. [#2545](https://github.com/zowe/zowe-cli/pull/2545)